#49 Move utility definition
