+++ necst/utils/__init__.py
@@ -1,2 +1,3 @@
 from .callbacks import *  # noqa: F401, F403
 from .ros_interface import *  # noqa: F401, F403
+from .ros_thread import *  # noqa: F401, F403
