+++ tests/core/test_commander.py
@@ -4,8 +4,9 @@ import pytest
 
 from necst import namespace, qos
 from necst.core import Authorizer, Commander
-from necst_msgs.msg import CoordMsg
-from ..conftest import TesterNode, destroy, spinning
+from necst.utils import spinning
+from necst_msgs.msg import CoordMsg, TimedAzElFloat64
+from ..conftest import TesterNode, destroy
 
 
 class TestCommander(TesterNode):
@@ -55,57 +56,48 @@ class TestCommander(TesterNode):
                 assert time.time() < timelimit, "Coordinate command not published in 2s"
                 time.sleep(0.02)
 
+        com.quit_privilege()
         destroy([com, auth_server])
         destroy(sub, self.node)
 
-    @pytest.mark.skip(reason="Method not established")
-    def test_antenna_point_with_wait(self):
-        ...
+    # def test_antenna_point_with_wait(self):
+    #     com = Commander()
+    #     auth_server = Authorizer()
+
+    #     cmd = {"lon": 30.0, "lat": 45.0, "unit": "deg", "frame": "altaz"}
 
     def test_antenna_stop(self):
         com = Commander()
         auth_server = Authorizer()
 
-        cmd = {
-            "lon": 30.0,
-            "lat": 45.0,
-            "unit": "deg",
-            "frame": "fk5",
-            "time": time.time(),
-        }
-        enc = {"lon": 10.0, "lat": 15.0, "unit": "deg", "frame": "altaz"}
-        checked = False
-
-        def check(msg: CoordMsg) -> None:
-            nonlocal checked
-            assert msg.lon == enc["lon"]
-            assert msg.lat == enc["lat"]
-            assert msg.unit == enc["unit"]
-            assert msg.frame == enc["frame"]
-            checked = True
-
         ns = namespace.antenna
-        sub = self.node.create_subscription(
-            CoordMsg, f"{ns}/raw_coord", check, qos.reliable
+        pub_speed = self.node.create_publisher(
+            TimedAzElFloat64, f"{ns}/speed", qos.realtime
         )
-        pub_enc = self.node.create_publisher(CoordMsg, f"{ns}/encoder", qos.realtime)
-        timer = self.node.create_timer(0.01, lambda: pub_enc.publish(CoordMsg(**enc)))
 
-        timelimit = time.time() + 2
-        with spinning([com, self.node, auth_server]):
+        start = time.time()
+
+        def publish_speed():
+            decay_factor = 1 / (time.time() - start + 1) ** 5
+            msg = TimedAzElFloat64(az=1.6 * decay_factor, el=1e-5, time=time.time())
+            pub_speed.publish(msg)
+
+        timer = self.node.create_timer(0.1, publish_speed)
+
+        with spinning([auth_server, self.node]):
             com.get_privilege()
-            com.antenna("stop", **cmd)
 
-            while not checked:
-                assert time.time() < timelimit, "Coordinate command not published in 2s"
-                time.sleep(0.02)
+            com.antenna("stop")
+            assert com.parameters["speed"].az < 1e-3
+            assert com.parameters["speed"].el < 1e-3
 
+            com.antenna("stop", lon=30, lat=45, frame="altaz", unit="deg")
+            assert com.parameters["speed"].az < 1e-3
+            assert com.parameters["speed"].el < 1e-3
+
+        com.quit_privilege()
         destroy([com, auth_server])
-        destroy([sub, pub_enc, timer], self.node)
-
-    @pytest.mark.skip(reason="Method not established")
-    def test_antenna_stop_with_wait(self):
-        ...
+        destroy([pub_speed, timer], node=self.node)
 
     @pytest.mark.skip(reason="Method not established")
     def test_wait(self):
