+++ tests/ctrl/test_sim_devices.py
@@ -1,11 +1,10 @@
 import time
 
-import pytest
-
 from necst import qos
 from necst.ctrl.antenna.sim_devices import AntennaDeviceSimulator
+from necst.utils import spinning
 from necst_msgs.msg import CoordMsg, TimedAzElFloat64
-from ..conftest import TesterNode, destroy, spinning
+from ..conftest import TesterNode, destroy
 
 
 class TestAntennaDeviceSimulator(TesterNode):
